formatowanie kodu i zmiana nazwy klasy
