+++ src/main/java/edu/kis/vh/nursery/FIFORhymer.java
@@ -1,21 +1,21 @@
 package edu.kis.vh.nursery;
 
-public class FIFORhymer extends defaultCountingOutRhymer {
+public class FIFORhymer extends DefaultCountingOutRhymer {
 
-	public defaultCountingOutRhymer temp = new defaultCountingOutRhymer();
-	
-	@Override
-	public int countOut() {
-		while (!callCheck())
-			
-		temp.countIn(super.countOut());
-		
-		int ret = temp.countOut();
-		
-		while (!temp.callCheck())
-			
-		countIn(temp.countOut());
-		
-		return ret;
-	}
+    public DefaultCountingOutRhymer temp = new DefaultCountingOutRhymer();
+
+    @Override
+    public int countOut() {
+        while (!callCheck())
+
+            temp.countIn(super.countOut());
+
+        int ret = temp.countOut();
+
+        while (!temp.callCheck())
+
+            countIn(temp.countOut());
+
+        return ret;
+    }
 }
