+++ src/main/java/edu/kis/vh/nursery/factory/DefaultRhymersFactory.java
@@ -1,30 +1,30 @@
 package edu.kis.vh.nursery.factory;
 
-import edu.kis.vh.nursery.defaultCountingOutRhymer;
+import edu.kis.vh.nursery.DefaultCountingOutRhymer;
 import edu.kis.vh.nursery.FIFORhymer;
 import edu.kis.vh.nursery.HanoiRhymer;
-import edu.kis.vh.nursery.factory.Rhymersfactory;
 
 public class DefaultRhymersFactory implements Rhymersfactory {
 
-	@Override
-	public defaultCountingOutRhymer GetStandardRhymer() {
-		return new defaultCountingOutRhymer();
-	}
+    @Override
+    public DefaultCountingOutRhymer GetStandardRhymer() {
+        return new DefaultCountingOutRhymer();
+    }
 
-	@Override
-	public defaultCountingOutRhymer GetFalseRhymer() {
-		return new defaultCountingOutRhymer();
-	}
+    @Override
+    public DefaultCountingOutRhymer GetFalseRhymer() {
+        return new DefaultCountingOutRhymer();
+    }
 
-	@Override
-	public defaultCountingOutRhymer GetFIFORhymer() {
-		return new FIFORhymer();
-	}
+    @Override
+    public DefaultCountingOutRhymer GetFIFORhymer() {
+        return new FIFORhymer();
+    }
 
-	@Override
-	public defaultCountingOutRhymer GetHanoiRhymer() {
-		return new HanoiRhymer();
-	}
+    @Override
+    public DefaultCountingOutRhymer GetHanoiRhymer() {
+        return new HanoiRhymer();
+    }
 
 }
+//alt strzalka - przechodzi miedzy otwartymi kartami
